# perfmon.py: make command failed due to package issue