So added that package for rhel and suse

Signed-off-by: shirisha Ganta <shiganta@in.ibm.com>