@shirishaganta 2020-04-27
Before Patch:
-----------------
localhost:/home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf # avocado run perfmon.py 
/usr/lib/python3.6/site-packages/avocado_framework-78.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
JOB ID     : 5093497bd8048d32745d72e23a04bd6005f822e3
JOB LOG    : /root/avocado-fvt-wrapper/results/job-2020-04-27T21.04-5093497/job.log
 (1/1) perfmon.py:Perfmon.test: ERROR: Command 'make' failed.\nstdout: b"Compiling for 'powerpc' target\nCompiling for 'Linux' system\nmake[1]: Entering directory '/var/tmp/avocado_1nwklniu/avocado_job_vbrhzrzt/1-perfmon.py_Perfmon.test/lib'\ncc  -g -Wall -Werror -Wextra -Wno-unused-parameter -... (10.40 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 10.56 s
JOB HTML   : /root/avocado-fvt-wrapper/results/job-2020-04-27T21.04-5093497/results.html

After Patch:
---------------
localhost:/home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf # avocado run perfmon.py 
/usr/lib/python3.6/site-packages/avocado_framework-78.0-py3.6.egg/avocado/plugins/run.py:287: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast, keep-tmp and ignore-missing-references.
  FutureWarning)
JOB ID     : 08109234ff5c1c2930096155eb2999eecb2a42dc
JOB LOG    : /root/avocado-fvt-wrapper/results/job-2020-04-27T21.06-0810923/job.log
 (1/1) perfmon.py:Perfmon.test: PASS (24.90 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 25.04 s
JOB HTML   : /root/avocado-fvt-wrapper/results/job-2020-04-27T21.06-0810923/results.html